Crude restoration of searching functionality
Very basic searching is possible again. MySIAM has to be used as
storage engine for the Post table and currently only the Content
field can be searched. A

mysql> ALTER TABLE Post ADD FULLTEXT (content);

command has to be executed, before anythig works and functionality
such as ordering by date or title and pagination is currently broken.

Fixed:

The fulltext index is now generated through Silverstrupe upon rebuilding
the database (only for the content filed of posts)

The results can be sorted by date in addition to relevancy score (sorting
by title is still not possible)